+++ NanoVNASaver/Charts/__init__.py
@@ -15,8 +15,9 @@ from .Permeability import PermeabilityChart
 from .Phase import PhaseChart
 from .QFactor import QualityFactorChart
 from .RI import RealImaginaryChart
-from .RIShunt import RealImaginaryShuntChart
-from .RISeries import RealImaginarySeriesChart
+from .RIZ import RealImaginaryZChart
+from .RIZShunt import RealImaginaryZShuntChart
+from .RIZSeries import RealImaginaryZSeriesChart
 from .Smith import SmithChart
 from .SParam import SParameterChart
 from .TDR import TDRChart
